pyproject.toml
[project]
name = "file-process-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "google-cloud-storage>=3.9.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.41.0",
]
